python_backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
motor==3.3.2
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0
bcrypt==4.1.2
pydantic==2.5.0
pydantic[email]==2.5.0
python-multipart==0.0.6 